CR-285 Move enums into their own package following review
