+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/enumerations/CalculationStatus.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations
+
+enum class CalculationStatus {
+  PRELIMINARY
+}
