+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/enumerations/SentenceIdentificationTrack.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model
+package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations
 
 enum class SentenceIdentificationTrack {
   SDS_BEFORE_CJA_LASPO,
