+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/model/BookingCalculation.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model
 
+import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.SentenceType
 import java.time.LocalDate
 
 data class BookingCalculation(
