--- src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/model/CalculationStatus.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model
-
-enum class CalculationStatus {
-  PRELIMINARY
-}
