+++ src/test/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/model/BookingCalculationTest.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.SentenceType
 import java.time.LocalDate
 
 internal class BookingCalculationTest {
